Add files via upload
The Online Bookstore stands as a testament to my forward-thinking approach to web development, having been conceptualized and implemented over a decade ago, prior to the advent of cutting-edge technologies such as artificial intelligence (AI). Despite the absence of AI-driven functionalities, the project remains a shining example of innovative web development practices and user-centric design principles.